requirements.txt
fastapi==0.111.0
itsdangerous==2.2.0
python-dotenv==1.0.1
pytz==2024.1
SQLAlchemy==2.0.31
